Refactored and Redesigned

+++ app/src/main/java/com/amt/doineedit_firebase/appDB/Item.kt
@@ -4,10 +4,19 @@ import com.google.firebase.database.Exclude
 import com.google.firebase.database.IgnoreExtraProperties
 
 @IgnoreExtraProperties
-data class Item(val itemName: String="", val price:Float= 0.0F, val quantity:Int=1, val haveItem:Boolean = false){
-    var showMenu = false
+data class Item(
+    val itemName: String = "",
+    val price: Double = 0.0,
+    val quantity: Int = 1,
+    val haveItem: Boolean = false
+) {
+    var showMenu = false // to validate Menu shown or not
+
     @Exclude
     fun toMap(): Map<String, Any?> {
+        /* used to edit item on database
+        * Return in Map Type which can be converted to JSON structure
+        * which was used on Firebase Realtime Database*/
         return mapOf(
             "itemName" to itemName,
             "price" to price,
